Valide los id que reciben los routes para evitar errores de casteo. Ej: /pacientes/1 es un id de formato invalido y arrojaria error de cast

+++ controllers/pacienteController.js
@@ -1,4 +1,5 @@
 const Paciente = require("../models/Paciente");
+const mongoose = require("mongoose");
 
 exports.nuevoPaciente = async (req, res, next) => {
   const paciente = new Paciente(req.body);
@@ -23,8 +24,12 @@ exports.obtenerPacientes = async (req,res,next) => {
 
 exports.obtenerPaciente = async (req,res,next) => {
   try {
-    const paciente = await Paciente.findById(req.params.id);
-    res.json(paciente);
+    if(mongoose.Types.ObjectId.isValid(req.params.id)){
+      const paciente = await Paciente.findById(req.params.id);
+      res.json(paciente);
+    }else{
+      res.json({ error: "Paciente inexistente"});
+    }
   } catch (error) {
     next(error);
   }
@@ -32,10 +37,14 @@ exports.obtenerPaciente = async (req,res,next) => {
 
 exports.eliminarPaciente = async (req,res,next) => {
   try {
-    await Paciente.findByIdAndDelete(req.params.id);
-    res.json({
-      mensaje: "Paciente eliminado correctamente"
-    });
+    if(mongoose.Types.ObjectId.isValid(req.params.id)){
+      await Paciente.findByIdAndDelete(req.params.id);
+      res.json({
+        mensaje: "Paciente eliminado correctamente"
+      });
+    }else{
+      res.json({ error: "Paciente inexistente"});
+    }
   } catch (error) {
     next(error);
   }
@@ -43,8 +52,13 @@ exports.eliminarPaciente = async (req,res,next) => {
 
 exports.actualizarPaciente = async (req,res,next) => {
   try {
-    const paciente = await Paciente.findByIdAndUpdate(req.params.id,req.body,{new:true});
-    res.json(paciente);
+    if(mongoose.Types.ObjectId.isValid(req.params.id)){
+      const paciente = await Paciente.findByIdAndUpdate(req.params.id,req.body,{new:true});
+      res.json(paciente);
+    }else{
+      res.json({ error: "Paciente inexistente"});
+    }
+    
   } catch (error) {
     next(error);
   }
